refactor(sources): isolate Discord contracts
